# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tài liệu API cho BTL Nhóm 2",
        "description": "Đây là tài liệu API cho ứng dụng điểm danh/quản lý lịch dạy.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/attendances": {
            "get": {
                "tags": [
                    "Attendance"
                ],
                "summary": "Lấy DS Điểm danh (Chưa triển khai)",
                "operationId": "getAttendancesList",
                "responses": {
                    "200": {
                        "description": "Chưa triển khai"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Attendance"
                ],
                "summary": "Tạo Điểm danh (Chưa triển khai)",
                "operationId": "storeAttendance",
                "responses": {
                    "200": {
                        "description": "Chưa triển khai"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/attendances/{attendance}": {
            "get": {
                "tags": [
                    "Attendance"
                ],
                "summary": "Lấy 1 Điểm danh (Chưa triển khai)",
                "operationId": "getAttendanceById",
                "parameters": [
                    {
                        "name": "attendance",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chưa triển khai"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Attendance"
                ],
                "summary": "Cập nhật Điểm danh (Chưa triển khai)",
                "operationId": "updateAttendance",
                "parameters": [
                    {
                        "name": "attendance",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chưa triển khai"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Attendance"
                ],
                "summary": "Xóa Điểm danh (Chưa triển khai)",
                "operationId": "deleteAttendance",
                "parameters": [
                    {
                        "name": "attendance",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chưa triển khai"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/schedules/{schedule}/students-attendance": {
            "get": {
                "tags": [
                    "Attendance"
                ],
                "summary": "Lấy danh sách SV và trạng thái điểm danh của 1 lịch học",
                "operationId": "getStudentsAndAttendanceForSchedule",
                "parameters": [
                    {
                        "name": "schedule",
                        "in": "path",
                        "description": "ID của lịch học (schedule)",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công. Trả về mảng danh sách SV và status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "student_id": {
                                                "type": "integer"
                                            },
                                            "student_name": {
                                                "type": "string"
                                            },
                                            "status": {
                                                "type": "string",
                                                "enum": [
                                                    "present",
                                                    "absent",
                                                    "late"
                                                ]
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/attendances/bulk-save": {
            "post": {
                "tags": [
                    "Attendance"
                ],
                "summary": "Lưu điểm danh hàng loạt cho một lịch học",
                "operationId": "saveBulkAttendance",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "schedule_id",
                                    "attendances"
                                ],
                                "properties": {
                                    "schedule_id": {
                                        "description": "ID của lịch học",
                                        "type": "integer",
                                        "example": 15
                                    },
                                    "attendances": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "student_id",
                                                "status"
                                            ],
                                            "properties": {
                                                "student_id": {
                                                    "type": "integer",
                                                    "example": 101
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "enum": [
                                                        "present",
                                                        "absent",
                                                        "late"
                                                    ],
                                                    "example": "present"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lưu điểm danh thành công"
                    },
                    "422": {
                        "description": "Dữ liệu không hợp lệ"
                    },
                    "500": {
                        "description": "Lỗi server khi lưu"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Đăng nhập người dùng",
                "operationId": "loginUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "teacher@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Đăng nhập thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Email hoặc mật khẩu không chính xác"
                    },
                    "422": {
                        "description": "Dữ liệu không hợp lệ (Validation error)"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Sử dụng Bearer token (Sanctum) sau khi đăng nhập",
                "name": "Bearer Token",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}